Fix receive modal responsive
