+++ src/modals/ReceiveModal/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
+import styled from 'styled-components';
 
 import { capitalize, lang } from 'balance-common';
 
@@ -25,6 +26,11 @@ const reduxProps = ({ account }) => ({
   accountType: account.accountType,
 });
 
+const StyledCardContainer = styled.div`
+  max-width: 440px;
+  width: 100%;
+`;
+
 class ReceiveModal extends Component {
   static propTypes = {
     modalClose: PropTypes.func.isRequired,
@@ -38,23 +44,27 @@ class ReceiveModal extends Component {
 
   render = () => {
     return (
-      <Card background="lightGrey">
-        <StyledContainer>
-          <StyledJustifyContent>
-            <StyledSubTitle>
-              <StyledIcon color="grey" icon={arrowUp} rotation={180} />
-              {lang.t('modal.receive_title', {
-                walletName: capitalize(
-                  `${this.props.accountType}${lang.t('modal.default_wallet')}`,
-                ),
-              })}
-            </StyledSubTitle>
-            <Button onClick={this.onClose}>{lang.t('button.close')}</Button>
-          </StyledJustifyContent>
-          <StyledQRCodeDisplay data={this.props.accountAddress} />
-          <StyledCopyToClipboard text={this.props.accountAddress} />
-        </StyledContainer>
-      </Card>
+      <StyledCardContainer>
+        <Card background="lightGrey">
+          <StyledContainer>
+            <StyledJustifyContent>
+              <StyledSubTitle>
+                <StyledIcon color="grey" icon={arrowUp} rotation={180} />
+                {lang.t('modal.receive_title', {
+                  walletName: capitalize(
+                    `${this.props.accountType}${lang.t(
+                      'modal.default_wallet',
+                    )}`,
+                  ),
+                })}
+              </StyledSubTitle>
+              <Button onClick={this.onClose}>{lang.t('button.close')}</Button>
+            </StyledJustifyContent>
+            <StyledQRCodeDisplay data={this.props.accountAddress} />
+            <StyledCopyToClipboard text={this.props.accountAddress} />
+          </StyledContainer>
+        </Card>
+      </StyledCardContainer>
     );
   };
 }
